separated the car calculations from CarInfo to carActions
